requirements.txt
Flask==3.0.3
mysql-connector-python==8.4.0
gunicorn==20.1.0
Werkzeug>=3.0.0



